extract.py
from PIL import Image
from pytesseract import pytesseract

def extract_text():
    path_to_tesseract = r"C:\Program Files\Tesseract-OCR\tesseract.exe"
    image_path = r"test.png"

    img = Image.open(image_path)

    pytesseract.tesseract_cmd = path_to_tesseract

    text = pytesseract.image_to_string(img)

    return text[:-1]
