requirements.txt
capturer==3.0
humanfriendly==10.0
packaging==23.2
pillow==10.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
pytesseract==0.3.10